refactor: seperated game and presentational logic
